+++ lib/CodeBuilder/Domain/Prototype/Collection.php
@@ -6,27 +6,44 @@ use IteratorAggregate;
 use Countable;
 use ArrayIterator;
 use InvalidArgumentException;
+use Traversable;
 
+/**
+ * @template T
+ * @implements IteratorAggregate<T>
+ */
 abstract class Collection implements IteratorAggregate, Countable
 {
-    protected $items = [];
+    /**
+     * @var T[]
+     */
+    protected array $items = [];
 
+    /**
+     * @param T[] $items
+     */
     protected function __construct(array $items)
     {
         $this->items = $items;
     }
 
+    /**
+     * @return static
+     */
     public static function empty()
     {
+        /** @phpstan-ignore-next-line */
         return new static([]);
     }
-
     
-    public function getIterator(): ArrayIterator
+    public function getIterator(): Traversable
     {
         return new ArrayIterator($this->items);
     }
 
+    /**
+     * @param T $item
+     */
     public function isLast($item): bool
     {
         return end($this->items) === $item;
@@ -34,10 +51,16 @@ abstract class Collection implements IteratorAggregate, Countable
 
     /**
      * Return first
+     * @return T|null
      */
     public function first()
     {
-        return reset($this->items);
+        $first = reset($this->items);
+        if (false === $first) {
+            return null;
+        }
+
+        return $first;
     }
 
     
@@ -46,6 +69,9 @@ abstract class Collection implements IteratorAggregate, Countable
         return count($this->items);
     }
 
+    /**
+     * @return T
+     */
     public function get(string $name)
     {
         if (!isset($this->items[$name])) {
@@ -69,6 +95,9 @@ abstract class Collection implements IteratorAggregate, Countable
         return false;
     }
 
+    /**
+     * @return static<T>
+     */
     public function notIn(array $names): Collection
     {
         return new static(array_filter($this->items, function ($name) use ($names) {
@@ -76,6 +105,9 @@ abstract class Collection implements IteratorAggregate, Countable
         }, ARRAY_FILTER_USE_KEY));
     }
 
+    /**
+     * @return static<T>
+     */
     public function in(array $names): Collection
     {
         return new static(array_filter($this->items, function ($name) use ($names) {