+++ lib/CodeBuilder/Domain/Prototype/ExtendsInterfaces.php
@@ -2,6 +2,9 @@
 
 namespace Phpactor\CodeBuilder\Domain\Prototype;
 
+/**
+ * @extends Collection<Type>
+ */
 class ExtendsInterfaces extends Collection
 {
     public static function fromTypes(array $types)