+++ lib/CodeBuilder/Domain/Prototype/Interfaces.php
@@ -2,9 +2,12 @@
 
 namespace Phpactor\CodeBuilder\Domain\Prototype;
 
+/**
+ * @extends Collection<InterfacePrototype>
+ */
 class Interfaces extends Collection
 {
-    public static function fromInterfaces(array $interfaces)
+    public static function fromInterfaces(array $interfaces): self
     {
         return new static($interfaces);
     }