+++ lib/WorseReflection/Core/Inference/FunctionStub/ArrayMapStub.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Phpactor\WorseReflection\Core\Inference\FunctionStub;
+
+use Microsoft\PhpParser\Node\DelimitedList\ArgumentExpressionList;
+use Microsoft\PhpParser\Node\Expression\ArgumentExpression;
+use Phpactor\WorseReflection\Core\Inference\Frame;
+use Phpactor\WorseReflection\Core\Inference\FunctionStub;
+use Phpactor\WorseReflection\Core\Inference\NodeContext;
+use Phpactor\WorseReflection\Core\Inference\NodeContextResolver;
+use Phpactor\WorseReflection\Core\TypeFactory;
+
+class ArrayMapStub implements FunctionStub
+{
+    public function resolve(
+        NodeContextResolver $resolver,
+        Frame $frame,
+        NodeContext $context,
+        ArgumentExpressionList $node
+    ): NodeContext {
+        $args = [];
+        foreach ($node->getChildNodes() as $expression) {
+            if (!$expression instanceof ArgumentExpression) {
+                continue;
+            }
+
+            $args[] = $resolver->resolveNode($frame, $expression);
+        }
+
+        if (!isset($args[0]) || !isset($args[1])) {
+            return $context;
+        }
+
+        return $context->withType(TypeFactory::array($args[0]->type()));
+    }
+}