+++ lib/WorseReflection/Core/Inference/Resolver/AnonymousFunctionCreationExpressionResolver.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Phpactor\WorseReflection\Core\Inference\Resolver;
+
+use Microsoft\PhpParser\Node;
+use Microsoft\PhpParser\Node\Expression\AnonymousFunctionCreationExpression;
+use Phpactor\WorseReflection\Core\Inference\Frame;
+use Phpactor\WorseReflection\Core\Inference\NodeContext;
+use Phpactor\WorseReflection\Core\Inference\NodeContextFactory;
+use Phpactor\WorseReflection\Core\Inference\NodeContextResolver;
+use Phpactor\WorseReflection\Core\Inference\Resolver;
+use Phpactor\WorseReflection\Core\Util\NodeUtil;
+
+class AnonymousFunctionCreationExpressionResolver implements Resolver
+{
+    public function resolve(NodeContextResolver $resolver, Frame $frame, Node $node): NodeContext
+    {
+        assert($node instanceof AnonymousFunctionCreationExpression);
+        $type = NodeUtil::typeFromQualfiedNameLike(
+            $resolver->reflector(),
+            $node,
+            $node->returnTypeList
+        );
+
+        return NodeContextFactory::create(
+            $node->getText(),
+            $node->getStartPosition(),
+            $node->getEndPosition(),
+            [
+                'type' => $type,
+            ]
+        );
+    }
+}