+++ lib/WorseReflection/Core/Type.php
@@ -4,6 +4,7 @@ namespace Phpactor\WorseReflection\Core;
 
 use Closure;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionScope;
+use Phpactor\WorseReflection\Core\Type\ArrayType;
 use Phpactor\WorseReflection\Core\Type\ClassType;
 use Phpactor\WorseReflection\Core\Type\Generalizable;
 use Phpactor\WorseReflection\Core\Type\MissingType;
@@ -47,9 +48,23 @@ abstract class Type
         return $this instanceof ClassType;
     }
 
+    public function isArray(): bool
+    {
+        return $this instanceof ArrayType;
+    }
+
     public function isNullable(): bool
     {
         return $this instanceof NullableType;
+    }
+
+    public function addToUnion(Type $type): UnionType
+    {
+        if (!$this instanceof UnionType) {
+            return new UnionType($this, $type);
+        }
+
+        return $this->add($type);
     }
 
     public function isPrimitive(): bool
@@ -104,11 +119,17 @@ abstract class Type
      */
     public function generalize(): Type
     {
-        if ($this instanceof Generalizable) {
-            return $this->generalize();
-        }
+        return $this->map(fn (Type $type) => $type instanceof Generalizable ? $type->generalize() : $type);
+    }
 
-        return $this;
+    public function equals(Type $type): bool
+    {
+        return $this->__toString() === $type->__toString();
+    }
+
+    public function instanceof(Type $type): Trinary
+    {
+        return Trinary::fromBoolean($type->equals($this));
     }
 
     /**